#107 [feat] old 파일 제거 및 리팩토링
