+++ src/components/Modal/MemoCollection/Image/MemoCollectionImage.Style.ts
@@ -1,43 +1,63 @@
 import styled from '@emotion/styled';
 import { theme } from '@styles/theme';
 
-export const MemoCollectionImageListContainer = styled.div({
+const MemoCollectionImageContainer = styled.div({
   display: 'flex',
   flexDirection: 'column',
-  alignItems: 'start',
+  gap: '32px',
 });
 
-export const MemoCollectionImageListTitle = styled.div({
-  ...theme.font.label1normal.semibold,
-  margin: '0',
+const MemoCollectionImageListContainer = styled.div({
+  display: 'flex',
+  flexDirection: 'column',
+});
+
+const MemoCollectionImageListTitle = styled.p({
   padding: '12px 16px',
-  color: theme.color.label.neutral,
-  boxShadow: theme.shadow.normal,
   borderTopLeftRadius: theme.radius[12],
   borderTopRightRadius: theme.radius[12],
+  boxShadow: theme.shadow.normal,
+
+  ...theme.font.label1normal.semibold,
+  color: theme.color.label.neutral,
+  margin: '0',
+  marginRight: 'auto',
 });
 
-export const MemoCollectionImageListContents = styled.div({
+const MemoCollectionImageListContents = styled.div({
   display: 'grid',
-  gridTemplateColumns: 'repeat(3, 1fr)',
-  rowGap: '12px',
+  gridAutoFlow: 'row',
+  gridTemplateColumns: 'repeat(3, minmax(0, 1fr))',
   padding: '18px',
-  width: '100%',
-  boxSizing: 'border-box',
+  columnGap: '16px',
+  rowGap: '16px',
+
   boxShadow: theme.shadow.emphasize,
   borderRadius: theme.radius[12],
   borderTopLeftRadius: '0',
+  background: theme.color.background.normal,
 });
 
-export const MemoColectionImageItemContainer = styled.div({
-  width: '172px',
-  height: '172px',
+const MemoCollectionImageItemContainer = styled.div({
+  display: 'flex',
+  width: 'auto',
+  aspectRatio: '1 / 1',
+  borderRadius: theme.radius[8],
+  position: 'relative',
   overflow: 'hidden',
-  borderRadius: '8px',
 
-  ['img']: {
-    objectFit: 'cover',
+  ['>img']: {
+    position: 'absolute',
     width: '100%',
     height: '100%',
+    objectFit: 'cover',
   },
 });
+
+export {
+  MemoCollectionImageContainer,
+  MemoCollectionImageListContainer,
+  MemoCollectionImageListTitle,
+  MemoCollectionImageListContents,
+  MemoCollectionImageItemContainer,
+};
