+++ src/components/Modal/MemoCollection/Link/MemoCollectionLink.Style.ts
@@ -1,89 +1,69 @@
 import styled from '@emotion/styled';
 import { theme } from '@styles/theme';
 
-export const MemoCollectionLinkContainer = styled.div({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '32px',
-});
-
-export const MemoCollectionLinkContents = styled.div({
+const MemoCollectionLinkContainer = styled.div({
   display: 'grid',
-  gridTemplateColumns: 'repeat(2, 1fr)',
+  gridAutoFlow: 'row',
+  gridTemplateColumns: 'repeat(2, minmax(0, 1fr))',
   columnGap: '12px',
   rowGap: '24px',
-  width: '580px',
-  boxSizing: 'border-box',
-  borderRadius: theme.radius[12],
-  borderTopLeftRadius: '0',
 });
 
-export const MemoCollectionLinkItemContainer = styled.div({
-  width: 'auto',
-  overflow: 'hidden',
+const MemoCollectionLinkItemContainer = styled.div({
+  display: 'flex',
+  flexDirection: 'column',
   borderRadius: theme.radius[12],
   padding: '12px',
   boxShadow: theme.shadow.emphasize,
-  display: 'flex',
-  flexDirection: 'column',
   gap: '12px',
-  textAlign: 'left',
-  cursor: 'pointer',
 });
 
-export const MemoCollectionLinkItemThumbnailContainer = styled.div({
-  width: '100%',
-  height: '200px',
+const MemoCollectionLinkItemImageContainer = styled.div({
+  aspectRatio: '4 / 3',
   borderRadius: theme.radius[6],
   overflow: 'hidden',
   position: 'relative',
+  cursor: 'pointer',
 
-  ['img']: {
-    objectFit: 'cover',
+  ['>img']: {
     width: '100%',
     height: '100%',
+    position: 'absolute',
+    top: '0',
+    left: '0',
+    objectFit: 'cover',
+  },
+
+  ['>svg']: {
+    position: 'absolute',
+    background: theme.color.material.dimmer,
+    borderRadius: theme.radius[4],
+    right: '0',
+    margin: '10px',
+    display: 'none',
+
+    [':hover']: {
+      background: theme.color.label.alternative,
+    },
+  },
+
+  [':hover']: {
+    ['>svg']: {
+      display: 'block',
+    },
   },
 });
 
-export const MemoCollectionLinkItemCategory = styled.div(
-  ({ color }: { color: string }) => ({
-    ['span']: {
-      background: color,
-    },
-  }),
-  {
-    position: 'absolute',
-    background: 'white',
-    left: '8px',
-    top: '8px',
-    borderRadius: theme.radius[48],
-    padding: '6px 10px',
-    display: 'flex',
-    gap: '4px',
-    alignItems: 'center',
+const MemoCollectionLinkItemTextContainer = styled.div({
+  width: '100%',
+  overflow: 'hidden',
 
-    ...theme.font.caption1.medium,
-    color: theme.color.label.normal,
-    margin: '0',
-
-    ['span']: {
-      width: '6px',
-      height: '6px',
-      borderRadius: theme.radius[6],
-    },
-  },
-);
-
-export const MemoCollectionLinkItemContents = styled.div({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '2px',
-
-  ['p']: {
+  ['>p']: {
     margin: '0',
     overflow: 'hidden',
     whiteSpace: 'nowrap',
     textOverflow: 'ellipsis',
+    textAlign: 'left',
 
     ['&.title']: {
       ...theme.font.caption1.semibold,
@@ -92,7 +72,7 @@ export const MemoCollectionLinkItemContents = styled.div({
     ['&.desc']: {
       ...theme.font.caption2.medium,
       color: theme.color.label.assistive,
-      marginBottom: '2px',
+      margin: '2px 0px 4px',
     },
     ['&.url']: {
       ...theme.font.caption2.medium,
@@ -100,3 +80,54 @@ export const MemoCollectionLinkItemContents = styled.div({
     },
   },
 });
+
+const MemoCollectionLinkItemCategoryContainer = styled.div(
+  ({ color }: { color?: string }) => ({
+    ['>span']: {
+      background: color,
+    },
+
+    ['::before']: {
+      background: color,
+    },
+  }),
+  {
+    background: 'white',
+    padding: '6px 10px',
+    position: 'absolute',
+    margin: '8px',
+    display: 'flex',
+    alignItems: 'center',
+    gap: '4px',
+    borderRadius: theme.radius[48],
+    overflow: 'hidden',
+    zIndex: '1',
+
+    ...theme.font.caption1.medium,
+    color: theme.color.label.normal,
+
+    ['>span']: {
+      width: '6px',
+      aspectRatio: '1 / 1',
+      borderRadius: theme.radius.full,
+    },
+
+    ['::before']: {
+      content: '""',
+      position: 'absolute',
+      width: '100%',
+      height: '100%',
+      opacity: '0.1',
+      left: 0,
+      zIndex: '-1',
+    },
+  },
+);
+
+export {
+  MemoCollectionLinkContainer,
+  MemoCollectionLinkItemContainer,
+  MemoCollectionLinkItemImageContainer,
+  MemoCollectionLinkItemTextContainer,
+  MemoCollectionLinkItemCategoryContainer,
+};
